"Made the registration page, which seems more functional than the homepage."
"Only 3 Members details asked for."
"Wanted to add a feature to have minimum 2 members and maximum 4 members
and ask for the minimum 2 members details first and then add more member
details if there are more members. But that could not be achieved."

"Other remaining problems in this page:

1. Tried to add the same preloader, but did not work for some unknown
reason.
2. Gender fields seem to have some sort of problem where if I select
gender option for one member, the selected option for the other member
is removed. Will look into that in the future. Tired Now.
3. Github icon is getting some blue colour and I cannot understand why.
4. Could try for a better transparent box to make it look better."

"Other jobs remaining:
1. Writing a READ-ME file.
2. Linking homepage to Reg_Page
3. Footer for homepage, but not possible now due to time contraints"